strnstr + protos update

+++ libft/transit/strlen.c
@@ -6,13 +6,15 @@
 /*   By: fcaquard <fcaquard@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/04 12:34:52 by fcaquard          #+#    #+#             */
-/*   Updated: 2021/04/04 12:34:53 by fcaquard         ###   ########.fr       */
+/*   Updated: 2021/04/04 22:02:57 by fcaquard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *str)
+#include <stddef.h>
+
+size_t	ft_strlen(const char *str)
 {
-	int index;
+	size_t index;
 
 	index = 0;
 	while (str[index])
